requirements.txt
asgiref==3.5.0
boto3==1.24.94
botocore==1.27.94
cachetools==5.3.0
certifi==2022.9.24
charset-normalizer==3.0.1
distlib==0.3.6
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.9.0
django-storages==1.13.2
djangorestframework==3.13.1
filelock==3.8.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
npm==0.1.1
optional-django==0.1.0
Pillow==10.4.0
pipenv==2022.10.12
platformdirs==2.5.2
proto-plus==1.23.0
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2021.3
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
whitenoise==6.2.0
